# Spring/sesion2-relaciones/src/main/java/com/ecommerce/controller/CustomerController.java
package com.ecommerce.controller;

import com.ecommerce.entity.Customer;
import com.ecommerce.service.CustomerService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

/**
 * CONTROLLER: CustomerController
 * ===============================
 * Expone los endpoints REST para la gestión de clientes (CRUD).
 *
 * URLs disponibles:
 *   GET    /api/customers        → todos los clientes
 *   GET    /api/customers/{id}   → uno por id
 *   POST   /api/customers        → crear cliente
 *   PUT    /api/customers/{id}   → actualizar cliente
 *   DELETE /api/customers/{id}   → eliminar cliente
 */
@RestController
@RequestMapping("/api/customers")
@RequiredArgsConstructor
public class CustomerController {

    private final CustomerService customerService;

    // GET /api/customers
    @GetMapping
    public ResponseEntity<List<Customer>> getAll() {
        return ResponseEntity.ok(customerService.getAll());
    }

    // GET /api/customers/{id}
    @GetMapping("/{id}")
    public ResponseEntity<Customer> getById(@PathVariable Long id) {
        return ResponseEntity.ok(customerService.getById(id));
    }

    // POST /api/customers
    // Body: {"name":"Pedro Ruiz","email":"pedro@email.com","address":"Av. Sol 100"}
    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public Customer create(@Valid @RequestBody Customer customer) {
        return customerService.create(customer);
    }

    // PUT /api/customers/{id}
    @PutMapping("/{id}")
    public ResponseEntity<Customer> update(
            @PathVariable Long id,
            @Valid @RequestBody Customer customer) {
        return ResponseEntity.ok(customerService.update(id, customer));
    }

    // DELETE /api/customers/{id}
    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable Long id) {
        customerService.delete(id);
    }
}